requirements.txt
# API
fastapi==0.115.12
uvicorn==0.34.3
pydantic==2.11.5
pydantic-settings==2.9.1
python-dotenv==1.1.0
python-multipart==0.0.9

# Versión del proyecto: 0.2.0

# MongoDB
motor==3.7.1
pymongo==4.13.0

# Cloudinary
cloudinary==1.36.0



# Desarrollo
pre-commit==3.6.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
requests==2.31.0

# PDF Generation
reportlab==4.4.2
Pillow>=11.3.0
pytest-json-report>=1.5.0
